Create shiftable module, move shifting functionality from enigma to it
