+++ lib/enigma.rb
@@ -1,9 +1,15 @@
 require './lib/shift'
+require './module/shiftable'
 
 class Enigma
+  include Shiftable
 
   def initialize
     @letters_array = ("a".."z").to_a << " "
+  end
+
+  def todays_date
+    Date.today.strftime("%d%m%y")
   end
 
   def startup(key, date, shift)
@@ -24,51 +30,7 @@ class Enigma
     end
   end
 
-  def encrypt_string(message, shift)
-    accum = []
-    message.downcase.chars.each_with_index do |letter, position|
-      if !@letters_array.include?(letter)
-        accum << letter
-      elsif position == 0 || position % 4 == 0
-        place = (@letters_array.index(letter) + shift.a_shift) % 27
-        accum << @letters_array[place]
-      elsif position == 1 || position % 4 == 1
-        place = (@letters_array.index(letter) + shift.b_shift) % 27
-        accum << @letters_array[place]
-      elsif position == 2 || position % 4 == 2
-        place = (@letters_array.index(letter) + shift.c_shift) % 27
-        accum << @letters_array[place]
-      elsif position == 3 || position % 4 == 3
-        place = (@letters_array.index(letter) + shift.d_shift) % 27
-        accum << @letters_array[place]
-      end
-    end
-    accum.join
-  end
-
-  def decrypt_string(message, shift)
-    accum = []
-    message.chars.each_with_index do |letter, position|
-      if !@letters_array.include?(letter)
-        accum << letter
-      elsif position == 0 || position % 4 == 0
-        place = (@letters_array.index(letter) - shift.a_shift) % 27
-        accum << @letters_array[place]
-      elsif position == 1 || position % 4 == 1
-        place = (@letters_array.index(letter) - shift.b_shift) % 27
-        accum << @letters_array[place]
-      elsif position == 2 || position % 4 == 2
-        place = (@letters_array.index(letter) - shift.c_shift) % 27
-        accum << @letters_array[place]
-      elsif position == 3 || position % 4 == 3
-        place = (@letters_array.index(letter) - shift.d_shift) % 27
-        accum << @letters_array[place]
-      end
-    end
-    accum.join
-  end
-
-  def encrypt(message, key = Key.new, date = Date.today.strftime("%d%m%y"))
+  def encrypt(message, key = Key.new, date = todays_date)
     encrypted_message = {}
     shift = Shift.new
     startup(key, date, shift)
@@ -79,7 +41,7 @@ class Enigma
     encrypted_message
   end
 
-  def decrypt(message, key = Key.new, date = Date.today.strftime("%d%m%y"))
+  def decrypt(message, key = Key.new, date = todays_date)
     decrypted_message = {}
     shift = Shift.new
     startup(key, date, shift)
